pyproject.toml
[tool.poetry]
name = "astraly-contracts"
version = "0.1.0"
description = ""
authors = ["astraly"]

[tool.poetry.dependencies]
python = "^3.8"
ecdsa = "^0.17.0"
fastecdsa = "^2.2.1"
sympy = "^1.9"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
cairo-lang = "0.9.1"
cairo-nile = "0.8.0"
dill = "^0.3.4"
pytest-xdist = "^2.4.0"
pytest-order = "^1.0.1"
wheel = "^0.37.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = {version = "^21.9b0", allow-prereleases = true}
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
